--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/smira/go-xz
+
+go 1.15