1. Add go mod
2. Reduce the number of code by half
3. Small performance gains  

      master branch
      ```txt
      goos: linux
      goarch: amd64
      cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
      BenchmarkDecompress-16               356           3030541 ns/op           21605 B/op        128 allocs/op
      PASS
      ok      github.com/smira/go-xz      1.424s
      ```
      optimize branch
      ```txt
      goos: linux
      goarch: amd64
      pkg: github.com/smira/go-xz
      cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
      BenchmarkDecompress-16               358           3022090 ns/op           21710 B/op        132 allocs/op
      PASS
      ok      github.com/smira/go-xz  1.425s
      ```